refact: remove private class of bundle
